% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynthesizeCirce.R
\name{createConceptSet}
\alias{createConceptSet}
\title{Create an OHDSI Circe ConceptSet Expression}
\usage{
createConceptSet(id, name, concepts)
}
\arguments{
\item{id}{Integer concept set identifier.}

\item{name}{Character string descriptive name.}

\item{concepts}{Data frame or list of concept tuples (conceptId, conceptName, domainId).}
}
\value{
A list conforming to OHDSI Circe ConceptSet schema.
}
\description{
Create an OHDSI Circe ConceptSet Expression
}
