% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{exportCirceJson}
\alias{exportCirceJson}
\title{Export Circe List to Formatted JSON File}
\usage{
exportCirceJson(circeList, filePath)
}
\arguments{
\item{circeList}{A list object representing an OHDSI Circe cohort definition.}

\item{filePath}{Destination file path.}
}
\description{
Export Circe List to Formatted JSON File
}
